requirements.txt
appnope==0.1.0
astroid==2.3.0
autopep8==1.4.4
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
Django==2.2.6
idna==2.9
imageio==2.6.1
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
joblib==0.14.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
matplotlib==3.1.1
mccabe==0.6.1
networkx==2.3
numpy==1.17.2
pandas==0.25.1
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==3.0.2
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.0.3
pyzmq==18.1.1
requests==2.23.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
selenium==3.14.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
torchvision==0.4.1
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
ujson==1.35
urllib3==1.23
wcwidth==0.1.8
wrapt==1.11.2
